refactor(tests): one decoder reads a grading-parity trial fixture, for every tier that drives one (#774)
The `grading_parity` `trial.yaml` decoder lived inside
`tests/canonical/test_grading_substrate_parity.py`, so a second tier driving
the same pack would have had to write its own reading of those bytes — and
"the canonical claim and the container agree" would have been a statement
about two hand-written readings rather than about one.

`tests/utils/grading_parity_packs.py` now holds it, exporting
`FIXTURE_TIMESTAMP`, `TrialCase`, `load_case` and `wire_message`. The
canonical differential imports the four names; its 36 use sites are renamed
mechanically. Semantics, docstrings and rejection messages are unchanged, and
the decoder's own contract locks stay in the canonical module, now reading
the shared decoder.

The move is proven pure by count, not by diff read: both canonical parity
suites collect and pass exactly what they collected and passed before —
`test_grading_substrate_parity.py` 135/135 and
`test_transcript_substrate_parity.py` 25/25, no skips either side. Making
`load_case` raise reds 94 of those 135, which is what says the canonical
suite reads this module and not a copy of it.
